revert path changes
